attendance/android/requirements.txt
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
gevent==25.9.1
greenlet==3.3.1
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
pycparser==3.0
Pygments==2.19.2
PyYAML==6.0.3
requests==2.32.5
rich==14.3.1
urllib3==2.6.3
websocket-client==1.6.2
zope.event==6.1
zope.interface==8.2
